Implement image upload route
Images can now be uploaded in the application creation process and they be linked to the application. Implement CORS policy to the app so the front end can make request from a different URL.

Fix mistake in applications table where phone number was an integer. Update some json schemas to better structured data for the front end. Fix eslint errors.